# api/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyFlow API
  description: SurveyFlow backend API for creating, distributing, and analyzing surveys.
  version: 1.0.0
  contact:
    name: SurveyFlow
    email: api@surveyflow.io
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/surveys/{surveyId}/qr-code:
    get:
      operationId: getQRCode
      tags: [Distribution]
      summary: Get QR code
      description: Generate a QR code for a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: size
          in: query
          schema:
            type: integer
        - name: error_level
          in: query
          schema:
            type: string
            enum: [L, M, Q, H]
      responses:
        "200":
          description: QR code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QRCodeResponse"
  /api/v1/surveys/{surveyId}/embed-code:
    get:
      operationId: getEmbedCode
      tags: [Distribution]
      summary: Get embed code
      description: Get HTML/JS embed code for a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: mode
          in: query
          schema:
            type: string
            enum: [popup, embedded, fullpage]
        - name: trigger
          in: query
          schema:
            type: string
            enum: [auto, button, custom]
        - name: width
          in: query
          schema:
            type: string
        - name: height
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Embed code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedCodeResponse"
...
  /api/v1/email-lists:
    get:
      operationId: listEmailLists
      tags: [Distribution]
      summary: List email lists
      description: List email lists for the organization.
      responses:
        "200":
          description: Email list array
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EmailList"
    post:
      operationId: createEmailList
      tags: [Distribution]
      summary: Create email list
      description: Create a new email list with optional contacts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailListRequest"
      responses:
        "201":
          description: Email list created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailList"
  /api/v1/email-lists/{listId}:
    get:
      operationId: getEmailList
      tags: [Distribution]
      summary: Get email list
      description: Get email list details.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
    put:
      operationId: updateEmailList
      tags: [Distribution]
      summary: Update email list
      description: Update an email list.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
    delete:
      operationId: deleteEmailList
      tags: [Distribution]
      summary: Delete email list
      description: Delete an email list.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
  /api/v1/email-lists/{listId}/contacts:
    post:
      operationId: addContacts
      tags: [Distribution]
      summary: Add contacts
      description: Add contacts to an email list.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
  /api/v1/email-lists/{listId}/contacts/{contactId}:
    delete:
      operationId: removeContact
      tags: [Distribution]
      summary: Remove contact
      description: Remove a contact from an email list.
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
  /api/v1/surveys/{surveyId}/send-emails:
    post:
      operationId: sendSurveyEmails
      tags: [Distribution]
      summary: Send survey emails
      description: Send a survey to an email list.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailsRequest"
      responses:
        "200":
          description: Emails queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendEmailsResponse"
  /api/v1/surveys/{surveyId}/email/campaigns:
    get:
      operationId: listEmailCampaigns
      tags: [Distribution]
      summary: List email campaigns
      description: List email campaigns for a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SurveyId:
      name: surveyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    StubError:
      type: object
      properties:
        code:
          type: string
          example: NOT_IMPLEMENTED
        message:
          type: string
          example: endpoint not yet implemented
        endpoint:
          type: string
...
    EmailList:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
        contact_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEmailListRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contacts:
          type: array
          items:
            type: object
            required: [email]
            properties:
              email:
                type: string
                format: email
              first_name:
                type: string
              last_name:
                type: string
              metadata:
                type: object
                additionalProperties: true
    SendEmailsRequest:
      type: object
      required: [list_id, subject]
      properties:
        list_id:
          type: string
          format: uuid
        subject:
          type: string
        body:
          type: string
        schedule_at:
          type: string
          format: date-time
    SendEmailsResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            job_id:
              type: string
              format: uuid
            count:
              type: integer
            status:
              type: string
    QRCodeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            svg:
              type: string
            png_url:
              type: string
              nullable: true
    EmbedCodeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            html:
              type: string
            js:
              type: string
...